divide main file into classes
