+++ Dynamics/__init__.py
@@ -12,8 +12,12 @@ from pymol import plugins
 
 from version import VERSION
 import pymol_plugin_dynamics
-
-
+import GromacsInput
+import GromacsOutput
+import SimulationParameters
+import Vectors
+import MdpConfig
+import CalculationWindow
 def __init_plugin__(app=None):
     p_label = f"Dynamics Gromacs {VERSION}"
     plugins.addmenuitemqt(p_label, pymol_plugin_dynamics.init_function)
