+++ Dynamics/pymol_plugin_dynamics.py
@@ -51,7 +51,12 @@ except ModuleNotFoundError:
 # Plugin Version
 from version import VERSION
 plugin_ver = f" {VERSION}"
-
+import GromacsInput
+import GromacsOutput
+import SimulationParameters
+import Vectors
+import MdpConfig
+import CalculationWindow
 EM_INIT_CONFIG = """define = -DFLEX_SPC
 constraints = none
 integrator = steep
@@ -176,7 +181,7 @@ def init_function(travis_ci=False, gui_library="qt", parent=False):
     elif gmx_version[0:4] not in supported_gmx_versions:
         print("Warning. Unsupported GROMACS Version")
     if status[0] == "ok":
-        simulation_parameters = SimulationParameters()
+        simulation_parameters = SimulationParameters.SimulationParameters()
     else:
         simulation_parameters = False
     if not travis_ci:
@@ -184,839 +189,8 @@ def init_function(travis_ci=False, gui_library="qt", parent=False):
     return status, simulation_parameters
 
 
-class SimulationParameters:
-    gmx_output = ""
-    gmx_input = ""
-    vectors_prody = False
-    stop = False
-    project_name = "nothing"
-    progress = ""
-    em_file = ""
-    pr_file = ""
-    md_file = ""
-
-    def __init__(self):
-        self.gmx_output = GromacsOutput()
-        self.gmx_input = GromacsInput()
-        print("Found GROMACS VERSION {}".format(self.gmx_output.version))
-        if prody:
-            self.vectors_prody = Vectors()
-            print("ProDy correctly imported")
-        self.progress = ProgressStatus()
-
-    def create_cfg_files(self):
-        self.em_file, self.pr_file, self.md_file = create_config_files(self.project_name)
-
-    def change_stop_value(self, value):
-        if value:
-            self.stop = True
-        else:
-            self.stop = False
-
-    def change_project_name(self, name):
-        self.project_name = name
-        project_dir = get_project_dirs(self.project_name)
-        if not os.path.isdir(project_dir):
-            os.makedirs(project_dir)
-
-
-# This class is responsible for interface to GROMACS. It will read all important data from GROMACS tools.
-class GromacsOutput:
-    version = "GROMACS not found"
-    command = ""
-    force_list = []
-    water_list = []
-    group_list = []
-    restraints = []
-
-    def __init__(self):
-        # Remove garbage
-        dynamics_dir = get_dynamics_dir()
-        garbage_files = next(os.walk(dynamics_dir))[2]
-        for garbage in garbage_files:
-            if garbage[0] == "#":
-                os.remove(dynamics_dir + garbage)
-
-        gmx_exe, gmx_version, gmx_build_arch, gmx_on_cygwin = get_gromacs_exe_info()
-        self.version = gmx_version
-        self.command = gmx_exe
-
-        # Track current directiry and switch to dynamics_dir before invoking gmx...
-        current_dir = os.getcwd()
-        os.chdir(dynamics_dir)
-
-        self.init2()
-
-        # Switch back to current directory...
-        os.chdir(current_dir)
-
-    def init2(self):
-        print("Reading available force fields and water models")
-
-        fo = open("test_gromacs.pdb", "wb")
-        fo.write(b"ATOM      1  N   LYS     1      24.966  -0.646  22.314  1.00 32.74      1SRN  99\n")
-        fo.close()
-
-        gmx_stdin_file_path = "gromacs_stdin.txt"
-        fo = open(gmx_stdin_file_path, "w")
-        fo.write("1\n")
-        fo.write("1")
-        fo.close()
-
-        gmx_stdout_file_path = "test_gromacs.txt"
-
-        cmd = "{} pdb2gmx -f test_gromacs.pdb -o test_gromacs.gro -p test_gromacs.top".format(self.command)
-        execute_subprocess(cmd, gmx_stdin_file_path, gmx_stdout_file_path)
-        lista_gromacs = read_text_lines(gmx_stdout_file_path)
-
-        # Reading available force fields
-        force_start_line = 0
-        while lista_gromacs[force_start_line] != "Select the Force Field:\n":
-            force_start_line = force_start_line + 1
-        force_start_line = force_start_line + 2
-        force_end_line = force_start_line
-        while lista_gromacs[force_end_line] != "\n":
-            force_end_line = force_end_line + 1
-        force_list = lista_gromacs[force_start_line:force_end_line]
-        force_list2 = []
-        number = 1
-        for force in force_list:
-            force_list2.append([number, force[:-1]])
-            number = number + 1
-
-        self.force_list = force_list2
-
-        # Reading available water models
-        self.water_list = get_water_models_info(lista_gromacs)
-
-        print("Reading available groups")
-
-        gmx_stdin_file_path = "gromacs_stdin.txt"
-        fo = open(gmx_stdin_file_path, "w")
-        fo.write("1")
-        fo.close()
-
-        gmx_stdout_file_path = "test_gromacs.txt"
-
-        cmd = "{} trjconv -f  test_gromacs.pdb -s test_gromacs.pdb -o test_gromacs2.pdb".format(self.command)
-        execute_subprocess(cmd, gmx_stdin_file_path, gmx_stdout_file_path)
-
-        group_test_list = read_text_lines(gmx_stdout_file_path)
-
-        # Reading available groups
-        group_start_line = 0
-        while group_test_list[group_start_line] != "Will write pdb: Protein data bank file\n":
-            group_start_line = group_start_line + 1
-        group_start_line = group_start_line + 1
-        group_end_line = group_start_line + 1
-        while group_test_list[group_end_line][0:14] != "Select a group":
-            group_end_line = group_end_line + 1
-        group_list = group_test_list[group_start_line:group_end_line]
-        group_list2 = []
-        number = 0
-        for group in group_list:
-            group1 = group.split(' has')
-            group2 = group1[0].split('Group     ')
-            if len(group2) == 2:
-                group_list2.append([number, group2[1]])
-                number = number + 1
-
-        self.group_list = group_list2
-
-    # This function will update water list if force field is changed.
-    def water_update(self, force_number):
-        # Track current directiry and switch to dynamics_dir before invoking gmx...
-        current_dir = os.getcwd()
-        os.chdir(get_dynamics_dir())
-
-        print("Updating available water models")
-        gmx_stdin_file_path = "gromacs_stdin.txt"
-        fo = open(gmx_stdin_file_path, "w")
-        fo.write("%d\n" % force_number)
-        fo.write("1")
-        fo.close()
-
-        gmx_stdout_file_path = "test_gromacs.txt"
-
-        cmd = "{} pdb2gmx -f  test_gromacs.pdb -o test_gromacs.gro -p test_gromacs.top".format(self.command)
-        execute_subprocess(cmd, gmx_stdin_file_path, gmx_stdout_file_path)
-
-        lista_gromacs = read_text_lines(gmx_stdout_file_path)
-        self.water_list = get_water_models_info(lista_gromacs)
-
-        # Switch back to current directory...
-        os.chdir(current_dir)
-
-        # save_options()
-        return self.water_list
-
-    # This function will read atoms group for restraints for current molecule.
-    def restraints_index(self, project_name):
-        self.restraints = []
-        current_dir = os.getcwd()
-        os.chdir(get_project_dirs(project_name))
-
-        fo = open("gromacs_stdin.txt", "w")
-        fo.write("q")
-        fo.close()
-
-        cmd = "{} make_ndx -f {}.pdb -o index.ndx".format(self.command, project_name)
-        execute_subprocess(cmd, "gromacs_stdin.txt", "restraints.log")
-
-        index_list = read_text_lines("restraints.log")
-
-        index_position = 0
-        atoms = ""
-        for line in index_list:
-            if line[0] == "[":
-                self.restraints.append([])
-                self.restraints[index_position].append(line)
-                if index_position != 0:
-                    self.restraints[index_position - 1].append(atoms)
-                index_position = index_position + 1
-                atoms = ""
-            else:
-                atoms = atoms + line
-        self.restraints[index_position - 1].append(atoms)
-        os.chdir(current_dir)
-
-
-# This class is responsible for performing molecular dynamics simulation with GROMACS tools.
-class GromacsInput:
-    force = 1
-    water = 1
-    group = 1
-    box_type = "triclinic"
-    explicit = 1
-    # variable to choose heavy hydrogen
-    hydro = "noheavyh"
-    box_distance = "0.8"
-    box_density = "1000"
-    restraints_nr = 1
-    # four variables salt, positive, negative and neutral
-    neutrality = "neutral"
-    salt_conc = "0.15"
-    positive_ion = "NA"
-    negative_ion = "CL"
-    command_distinction = "\n!************************!\n"
-
-    # This function will change given variabless stored by the class (needed for lambda statements)
-    def update(self, gmx_options):
-        for key, value in gmx_options.items():
-            if key == "force":
-                self.force = value
-            elif key == "water":
-                self.water = value
-            elif key == "group":
-                self.group = value
-            elif key == "box_type":
-                self.box_type = value
-            elif key == "hydro":
-                self.hydro = value
-            elif key == "box_distance":
-                self.box_distance = value
-            elif key == "box_density":
-                self.box_density = value
-            elif key == "restraints_nr":
-                self.restraints_nr = value
-            elif key == "neutrality":
-                self.neutrality = value
-            elif key == "salt_conc":
-                self.salt_conc = value
-            elif key == "positive_ion":
-                self.positive_ion = value
-            elif key == "negative_ion":
-                self.negative_ion = value
-            elif key == "explicit":
-                self.explicit = value
-        #        save_options()
-        print("gromacs updated")
-
-    # This function will create initial topology and trajectory using pdb file and choosen force field
-    def pdb2top(self, s_params):
-        status = ["ok", "Calculating topology using Force fields"]
-        status_update(status)
-        hh = "-" + self.hydro
-        gmx_cmd = s_params.gmx_output.command
-        project_name = s_params.project_name
-        try:
-            os.remove("{}.gro".format(project_name))
-            os.remove("{}.top".format(project_name))
-        except FileNotFoundError:
-            pass
-
-        fo = open("gromacs_stdin.txt", "w")
-        fo.write("%s\n" % str(self.force))
-        fo.write("%s" % str(self.water))
-        fo.close()
-
-        command = "{0} pdb2gmx -f {1}.pdb -o {1}.gro -p {1}.top {2}".format(gmx_cmd, project_name, hh)
-        execute_and_monitor_subprocess(command, 'gromacs_stdin.txt', 'log1.txt', 'log.txt')
-
-        if os.path.isfile("{}.gro".format(project_name)):
-            status = ["ok", ""]
-        else:
-            status = ["fail", "Warning. Trying to ignore unnecessary hydrogen atoms."]
-
-            command = "{0} pdb2gmx -ignh -f {1}.pdb -o {1}.gro -p {1}.top {2}".format(gmx_cmd, project_name, hh)
-            execute_and_monitor_subprocess(command, 'gromacs_stdin.txt', 'log1.txt', 'log.txt')
-        status_update(status)
-
-        stop = s_params.stop
-        if os.path.isfile("{}.gro".format(project_name)) and not stop:
-            status = ["ok", "Calculated topology using Force fields"]
-        else:
-            status = ["fail", "Force field unable to create topology file"]
-        return status
-
-    # This is alternative function to create initial topology and triectory using pdb file
-    @staticmethod
-    def x2top(s_params):
-        status = ["ok", "Calculating topology using Force fields"]
-        status_update(status)
-        gmx_cmd = s_params.gmx_output.command
-        project_name = s_params.project_name
-        try:
-            os.remove("{}.gro".format(project_name))
-            os.remove("{}.top".format(project_name))
-        except FileNotFoundError:
-            pass
-
-        command = "{0} x2top -f {1}.pdb -o {1}.top".format(gmx_cmd, project_name)
-        execute_and_monitor_subprocess(command, None, 'log1.txt', 'log.txt')
-
-        stop = s_params.stop
-        if os.path.isfile("{}.top".format(project_name)) and not stop:
-            status = ["ok", "Calculating structure using trjconv."]
-        else:
-            status = ["fail", "Unable to create topology file."]
-        status_update(status)
-
-        if status[0] == "ok":
-            fo = open("gromacs_stdin.txt", "w")
-            fo.write("0")
-            fo.close()
-
-            command = "{0} trjconv -f {1}.pdb -s {1}.pdb -o {1}.gro".format(gmx_cmd, project_name)
-            execute_and_monitor_subprocess(command, 'gromacs_stdin.txt', 'log1.txt', 'log.txt')
-
-            stop = s_params.stop
-            if os.path.isfile("{}.gro".format(project_name)) and not stop:
-                status = ["ok", "Calculated structure using trjconv."]
-            else:
-                status = ["fail", "Unable to create structure file."]
-        return status
-
-    # This function will create and add waterbox.
-    def waterbox(self, s_params):
-        status = ["ok", "Generating waterbox"]
-        box_type = "-bt {}".format(self.box_type)
-        distance = "-d {}".format(self.box_distance)
-        density = "-density {}".format(self.box_density)
-        gmx_cmd = s_params.gmx_output.command
-        project_name = s_params.project_name
-        try:
-            os.remove("{}1.gro".format(project_name))
-            os.remove("{}_solv.gro".format(project_name))
-        except FileNotFoundError:
-            pass
-
-        status_update(status)
-        command = "{0} editconf -f {1}.gro -o {1}1.gro -c {2} {3} {4}".format(gmx_cmd, project_name, box_type, distance,
-                                                                              density)
-        execute_and_monitor_subprocess(command, None, 'log1.txt', 'log.txt')
-
-        water_name = s_params.gmx_output.water_list[self.water - 1][1][4:8].lower()
-        print(water_name)
-        if water_name == "tip4":
-            water_gro = "tip4p.gro"
-        elif water_name == "tip5":
-            water_gro = "tip5p.gro"
-        else:
-            water_gro = "spc216.gro"
-
-        command = "{0} solvate -cp {1}1.gro -cs {2} -o {1}_solv.gro -p {1}.top".format(gmx_cmd, project_name, water_gro)
-
-        status = ["ok", "Adding Water Box"]
-        status_update(status)
-
-        execute_and_monitor_subprocess(command, None, 'log1.txt', 'log.txt')
-
-        stop = s_params.stop
-        if os.path.isfile("{}1.gro".format(project_name)) and not stop:
-            status = ["ok", "Water Box Added"]
-        else:
-            status = ["fail", "Unable to add water box"]
-        return status
-
-    # This function will add ions/salts to the protein in waterbox
-    def saltadd(self, s_params):
-        status = ["ok", "Preparing to add ions or salt"]
-        salt = "-conc {}".format(self.salt_conc)
-        positive = "-pname {}".format(self.positive_ion)
-        negative = "-nname {}".format(self.negative_ion)
-        neu = "-{}".format(self.neutrality)
-        gmx_cmd = s_params.gmx_output.command
-        project_name = s_params.project_name
-        try:
-            os.remove(project_name + "_b4em.gro")
-            os.remove(project_name + "_ions.tpr")
-        except FileNotFoundError:
-            pass
-
-        command = "{0} grompp -f em -c {1}_solv.gro -o {1}_ions.tpr -p {1}.top -maxwarn 1".format(gmx_cmd, project_name)
-        status_update(status)
-        execute_and_monitor_subprocess(command, None, 'log1.txt', 'log.txt')
-
-        fo = open("gromacs_stdin.txt", "w")
-        fo.write("13")
-        fo.close()
-
-        status = ["ok", "Adding salts and ions"]
-        status_update(status)
-
-        command = "{0} genion -s {1}_ions.tpr -o {1}_b4em.gro {2} {3} {4} {5} -p {1}.top".format(gmx_cmd,
-                                                                                                 project_name, positive,
-                                                                                                 negative, salt, neu)
-        execute_and_monitor_subprocess(command, 'gromacs_stdin.txt', 'log1.txt', 'log.txt')
-
-        stop = s_params.stop
-        if os.path.isfile("{}_b4em.gro".format(project_name)) and not stop:
-            status = ["ok", "Ions added successfully"]
-        elif stop == 0:
-            status = ["ok", "Find out what's wrong!"]
-        else:
-            status = ["failed", "Unable to add ions"]
-        return status
-
-    # This function will perform energy minimization
-    @staticmethod
-    def em(s_params):
-        status = ["ok", "Energy Minimization"]
-        gmx_cmd = s_params.gmx_output.command
-        project_name = s_params.project_name
-        try:
-            os.remove("{}_em.tpr".format(project_name))
-            os.remove("{}_em.trr".format(project_name))
-            os.remove("{}_b4pr.gro".format(project_name))
-        except FileNotFoundError:
-            pass
-
-        # Check if waterbox was added and adjust accordingly.
-        if not os.path.isfile("{}_b4em.gro".format(project_name)):
-            if os.path.isfile("{}_solv.gro".format(project_name)):
-                shutil.copy("{}_solv.gro".format(project_name), "{}_b4em.gro".format(project_name))
-            elif os.path.isfile(project_name + "{}.gro".format(project_name)):
-                shutil.copy("{}.gro".format(project_name), "{}_b4em.gro".format(project_name))
-
-        status_update(status)
-        command = "{0} grompp -f em -c {1}_b4em -p {1} -o {1}_em".format(gmx_cmd, project_name)
-        execute_and_monitor_subprocess(command, None, 'log1.txt', 'log.txt')
-
-        command = "{0} mdrun -nice 4 -s {1}_em -o {1}_em -c {1}_b4pr -v".format(gmx_cmd, project_name)
-        execute_and_monitor_subprocess(command, None, 'log1.txt', 'log.txt')
-
-        stop = s_params.stop
-        if os.path.isfile("{}_em.tpr".format(project_name)) and os.path.isfile("{}_b4pr.gro".format(project_name)) and \
-                not stop:
-            status = ["ok", "Energy Minimized"]
-        else:
-            status = ["fail", "Unable to perform Energy Minimization"]
-        return status
-
-    # This function will perform position restrained MD
-    @staticmethod
-    def pr(s_params):
-        status = ["ok", "Position Restrained MD"]
-        gmx_cmd = s_params.gmx_output.command
-        project_name = s_params.project_name
-        try:
-            os.remove("{}_pr.tpr".format(project_name))
-            os.remove("{}_pr.trr".format(project_name))
-            os.remove("{}_b4md.gro".format(project_name))
-        except FileNotFoundError:
-            pass
-
-        status_update(status)
-        command = "{0} grompp -f pr -c {1}_b4pr -r {1}_b4pr -p {1} -o {1}_pr".format(gmx_cmd, project_name)
-        execute_and_monitor_subprocess(command, None, 'log1.txt', 'log.txt')
-
-        command = "{0} mdrun -nice 4 -s {1}_pr -o {1}_pr -c {1}_b4md -v".format(gmx_cmd, project_name)
-        execute_and_monitor_subprocess(command, None, 'log1.txt', 'log.txt')
-
-        stop = s_params.stop
-        if os.path.isfile("{}_pr.tpr".format(project_name)) and not stop:
-            status = ["ok", "Position Restrained MD finished"]
-        else:
-            status = ["fail", "Unable to perform Position Restrained"]
-        return status
-
-    # This function will create posre.itp file for molecular dynamics simulation with choosen atoms if
-    # restraints were selected
-    @staticmethod
-    def restraints(s_params):
-        status = ["ok", "Adding Restraints"]
-        gmx_cmd = s_params.gmx_output.command
-        project_name = s_params.project_name
-        try:
-            os.remove("posre_2.itp")
-        except FileNotFoundError:
-            pass
-
-        fo = open("gromacs_stdin.txt", "w")
-        fo.write("0")
-        fo.close()
-
-        status_update(status)
-        command = "{0} genrestr -f {1}.pdb -o posre_2.itp -n index_dynamics.ndx".format(gmx_cmd, project_name)
-        execute_and_monitor_subprocess(command, 'gromacs_stdin.txt', 'log1.txt', 'log.txt')
-
-        stop = s_params.stop
-        if os.path.isfile("posre_2.itp") and not stop:
-            status = ["ok", "Added Restraints"]
-            if os.path.isfile("posre.itp"):
-                os.remove("posre.itp")
-            shutil.copy("posre_2.itp", "posre.itp")
-        else:
-            status = ["fail", "Unable to create restraints file"]
-        return status
-
-    # This function will perform position final molecular dynamics simulation
-    @staticmethod
-    def md(s_params):
-        status = ["ok", "Molecular Dynamics Simulation"]
-        gmx_cmd = s_params.gmx_output.command
-        project_name = s_params.project_name
-        try:
-            os.remove("{}_md.tpr".format(project_name))
-            os.remove("{}_md.trr".format(project_name))
-        except FileNotFoundError:
-            pass
-
-        # Check if em and/or pr was done and adjust accordingly.
-        if not os.path.isfile("{}_b4md.gro".format(project_name)):
-            if not os.path.isfile("{}_b4pr.gro".format(project_name)):
-                # No em and pr
-                shutil.copy("{}_b4em.gro".format(project_name), "{}_b4md.gro".format(project_name))
-            else:
-                # No pr
-                shutil.copy("{}_b4pr.gro".format(project_name), "{}_b4md.gro".format(project_name))
-
-        status_update(status)
-        command = "{0} grompp -f md -c {1}_b4md  -p {1} -o {1}_md".format(gmx_cmd, project_name)
-        execute_and_monitor_subprocess(command, None, 'log1.txt', 'log.txt')
-
-        command = "{0} mdrun -nice 4 -s {1}_md -o {1}_md -c {1}_after_md -v".format(gmx_cmd, project_name)
-        execute_and_monitor_subprocess(command, None, 'log1.txt', 'log.txt')
-
-        stop = s_params.stop
-        if os.path.isfile("{}_md.tpr".format(project_name)) and not stop:
-            status = ["ok", "Molecular Dynamics Simulation finished"]
-        else:
-            status = ["fail", "Unable to perform Molecular Dynamics Simulation"]
-        return status
-
-    # This function will convert final results to multimodel pdb file
-    def trjconv(self, s_params):
-        status = ["ok", "Creating Multimodel PDB"]
-        gmx_cmd = s_params.gmx_output.command
-        project_name = s_params.project_name
-        try:
-            os.remove("{}_multimodel.pdb".format(project_name))
-        except FileNotFoundError:
-            pass
-        if os.path.isfile("{}_multimodel.pdb".format(project_name)):
-            os.remove("{}_multimodel.pdb".format(project_name))
-
-        fo = open("gromacs_stdin.txt", "w")
-        fo.write("%s" % str(self.group))
-        fo.close()
-
-        status_update(status)
-        command = "{0} trjconv -f {1}_md.trr -s {1}_md.tpr -o {1}_multimodel.pdb".format(gmx_cmd, project_name)
-        execute_and_monitor_subprocess(command, 'gromacs_stdin.txt', 'log1.txt', 'log.txt')
-
-        stop = s_params.stop
-        if os.path.isfile("{}_multimodel.pdb".format(project_name)) and not stop:
-            status = ["ok", "Finished!"]
-        else:
-            status = ["fail", "Unable to generate multimodel PDB file"]
-        return status
-
-
-# This class will handle PCA by ProDy python library and show vectors from NMD file.
-class Vectors:
-    nmd_name = []
-    nmd_atomnames = []
-    nmd_resnames = []
-    nmd_resids = []
-    nmd_bfactors = []
-    nmd_coordinates = []
-    nmd_mode = []
-    nmd_scale_mode = []
-    color = "grey"
-    scale = 1.0
-    mode_nr = 0
-
-    calculation_type = 0
-    contact_map = 0
-    block_contact_map = 0
-
-    enm = 0
-
-    # Change Multimodel PDB file into NMD vector file
-    def prody(self, project_name):
-        # Silence ProDy and create logs
-        prody.confProDy(verbosity='none')
-        prody.startLogfile("log_prody.log")
-        # Prepare ensemble
-        model = prody.parsePDB(project_name + "_multimodel.pdb", subset='calpha')
-        ensemble = prody.Ensemble(project_name + ' ensemble')
-        ensemble.setCoords(model.getCoords())
-        ensemble.addCoordset(model.getCoordsets())
-        ensemble.iterpose()
-        # ANM calculations
-        if self.calculation_type == 0:
-            anm = prody.ANM(project_name)
-            anm.buildHessian(ensemble)
-            anm.calcModes()
-            write_nmd = anm
-            self.enm = anm
-        # PCA calculations
-        elif self.calculation_type == 1:
-            pca = prody.PCA(project_name)
-            pca.buildCovariance(ensemble)
-            pca.calcModes()
-            write_nmd = pca
-        # GNM calculations
-        elif self.calculation_type == 2:
-            gnm = prody.GNM(project_name)
-            gnm.buildKirchhoff(ensemble)
-            gnm.calcModes()
-            write_nmd = gnm
-            self.enm = gnm
-        # Write NMD file
-        prody.writeNMD(project_name + '.nmd', write_nmd[:3], model)
-        prody.closeLogfile("log_prody.log")
-
-    # Read NMD file
-    def nmd_format(self, project_name):
-        file_nmd = open('{}.nmd'.format(project_name), "r")
-        list_nmd = file_nmd.readlines()
-
-        self.nmd_mode = []
-        self.nmd_scale_mode = []
-        for line in list_nmd:
-            split_line = line.split()
-            if split_line[0] == "name":
-                self.nmd_name = split_line
-                self.nmd_name.pop(0)
-            elif split_line[0] == "atomnames":
-                self.nmd_atomnames = split_line
-                self.nmd_atomnames.pop(0)
-            elif split_line[0] == "resnames":
-                self.nmd_resnames = split_line
-                self.nmd_resnames.pop(0)
-            elif split_line[0] == "resids":
-                self.nmd_resids = split_line
-                self.nmd_resids.pop(0)
-            elif split_line[0] == "bfactors":
-                self.nmd_bfactors = split_line
-                self.nmd_bfactors.pop(0)
-            elif split_line[0] == "coordinates":
-                self.nmd_coordinates = split_line
-                self.nmd_coordinates.pop(0)
-            elif split_line[0] == "mode":
-                pre_mode = split_line
-                self.nmd_mode.append(pre_mode[3:])
-                self.nmd_scale_mode.append(pre_mode[2])
-
-    # Show contact map on PyMOL screen
-    def show_contact_map(self, sensitivity, project_name):
-        contact_matrix = self.enm.getKirchhoff()
-        print(contact_matrix)
-        c_alpha_nr = 0
-        for c_alpha_list in contact_matrix:
-            c_alpha_nr = c_alpha_nr + 1
-            c_alpha_target_nr = 0
-            for c_alpha_1 in c_alpha_list:
-                c_alpha_target_nr = c_alpha_target_nr + 1
-                if c_alpha_nr != c_alpha_target_nr and float(c_alpha_1) < float(sensitivity):
-                    cmd.select("sele1",
-                               "n. ca and {}_multimodel and i. {}".format(project_name, str(c_alpha_nr)))  # PyMOL API
-                    cmd.select("sele2", "n. ca and {}_multimodel and i. {}".format(project_name,
-                                                                                   str(c_alpha_target_nr)))  # PyMOL API
-                    cmd.distance("contact_map", "sele1", "sele2")  # PyMOL API
-        try:
-            cmd.hide("labels", "contact_map")  # PyMOL API
-            cmd.delete("sele1")  # PyMOL API
-            cmd.delete("sele2")  # PyMOL API
-        except:
-            pass
-
-    # Show contact map/cross corelation as a graph
-    def graph_contact_map(self, plot_type):
-        if plot_type == "contact":
-            # matplotlib
-            prody.showContactMap(self.enm)
-        elif plot_type == "cross":
-            # matplotlib
-            prody.showCrossCorr(self.enm)
-
-    # Show vectors from NMD file
-    def show_vectors(self):
-        color1 = cmd.get_color_tuple(self.color)  # PyMOL API
-        color2 = cmd.get_color_tuple(self.color)  # PyMOL API
-        if color1:
-            color1 = list(color1)
-        # Fallback to grey in case of unrecognized color
-        else:
-            color1 = [0.5, 0.5, 0.5]
-        if color2:
-            color2 = list(color2)
-        # Fallback to grey in case of unrecognized color
-        else:
-            color2 = [0.5, 0.5, 0.5]
-        arrow_head_radius = 0.15
-
-        x1 = []
-        y1 = []
-        z1 = []
-
-        coor = "x"
-        for coordinate in self.nmd_coordinates:
-            if coor == "x":
-                x1.append(float(coordinate))
-                coor = "y"
-            elif coor == "y":
-                y1.append(float(coordinate))
-                coor = "z"
-            elif coor == "z":
-                z1.append(float(coordinate))
-                coor = "x"
-
-        x2 = []
-        y2 = []
-        z2 = []
-
-        # This factor is provided to make vector length more like in NMWiz.
-        # More investigation is needed to get exact formula.
-        approximation_factor = 16.6
-
-        coor = "x"
-        coor_nr = 0
-        round_nr = 0
-        for mode in self.nmd_mode[self.mode_nr]:
-            if coor == "x":
-                x2.append(
-                    float(mode) * float(self.nmd_scale_mode[self.mode_nr]) * approximation_factor * self.scale + x1[
-                        coor_nr])
-                coor = "y"
-            elif coor == "y":
-                y2.append(
-                    float(mode) * float(self.nmd_scale_mode[self.mode_nr]) * approximation_factor * self.scale + y1[
-                        coor_nr])
-                coor = "z"
-            elif coor == "z":
-                z2.append(
-                    float(mode) * float(self.nmd_scale_mode[self.mode_nr]) * approximation_factor * self.scale + z1[
-                        coor_nr])
-                coor = "x"
-            round_nr = round_nr + 1
-            if round_nr == 3:
-                round_nr = 0
-                coor_nr = coor_nr + 1
-
-        coor_nr = 0
-        for position in x1:
-            try:
-                cmd.delete("Mode_Vector_" + str(coor_nr))
-            except:
-                pass
-
-            cone = [cgo.CONE, x1[coor_nr], y1[coor_nr], z1[coor_nr], x2[coor_nr], y2[coor_nr], z2[coor_nr],
-                    arrow_head_radius, 0.0] + color1 + color2 + [1.0, 0.0]
-            cmd.load_cgo(cone, "Mode_Vector_" + str(coor_nr))  # PyMOL API
-            coor_nr = coor_nr + 1
-        # Another workaround for PyMOL 1.8 with TravisCI
-        try:
-            cam_possition = cmd.get_view(quiet=1)  # PyMOL API
-            cmd.set_view(cam_possition)  # PyMOL API
-        except TypeError:
-            pass
-
-    def change_vectors_color(self, color):
-        self.color = color
-        self.show_vectors()
-
-    def change_vectors_scale(self, scale):
-        scale = float(scale)
-        self.scale = scale
-        self.show_vectors()
-
-    def change_vectors_mode_nr(self, mode_nr):
-        self.mode_nr = mode_nr
-        self.show_vectors()
-
-    def options_change(self, v1, v2, root):
-        self.calculation_type = v1.get()
-        self.contact_map = v2.get()
-        # save_options()
-        root.destroy()
-
-    def block_contact(self, block, contact_map_b, contact_map_v):
-        # ToDO: Replace below Tk code with something agnostic
-        self.block_contact_map = block
-        if block == 0:
-            contact_map_b.configure(state=ACTIVE)
-        elif block == 1:
-            contact_map_b.configure(state=DISABLED)
-            contact_map_v.set(0)
-
 
 # This class create and maintain abstraction mdp file representatives. em.mdp, pr.mdp, md.mdp
-class MdpConfig:
-    external_file = 0
-    options = [[]]
-    file_name = ""
-
-    def __init__(self, file_name, init_config, external_file=0):
-        self.file_name = file_name
-        self.external_file = external_file
-        list1 = init_config.split("\n")
-        list2 = []
-        for line in list1:
-            list2.append(line.split(" = "))
-        self.options = list2
-
-    def update(self, option_nr, value, check=1):
-        self.options[option_nr][1] = value
-        if check == 0 and self.options[option_nr][0][0] != ";":
-            self.options[option_nr][0] = ";" + self.options[option_nr][0]
-        elif check == 1 and self.options[option_nr][0][0] == ";":
-            self.options[option_nr][0] = self.options[option_nr][0][1:]
-        self.clean_artefacts()
-
-    def save_file(self, s_params):
-        project_name = s_params.project_name
-        project_dir = get_project_dirs(project_name)
-        config = ""
-        for option in self.options:
-            # pass empty option
-            if option == ['']:
-                pass
-            else:
-                config = "{}{} = {}\n".format(config, str(option[0]), str(option[1]))
-        mdp = open(project_dir + self.file_name, "w")
-        mdp.write(config)
-        mdp.close()
-
-    # Clean options from artefacts
-    def clean_artefacts(self):
-        try:
-            self.options.remove([''])
-        except:
-            pass
 
 
 # This function creates files needed by the project
@@ -1256,6 +430,8 @@ def get_water_models_info(gmx_output_lines):
         end_line = end_line + 1
 
     waters_info = gmx_output_lines[start_line:end_line]
+    if( len(waters_info)<=0):
+        waters_info = get_alternate_water_info(gmx_output_lines)
     waters_info2 = []
     number = 1
     for water in waters_info:
@@ -1263,6 +439,16 @@ def get_water_models_info(gmx_output_lines):
         number = number + 1
 
     return waters_info2
+#In newwer version of Gromacs water is under Section Select Water Model
+def get_alternate_water_info(gmx_output_lines):
+    wm_start_line = 0
+    while gmx_output_lines[wm_start_line] != "Select the Water Model:\n":
+            wm_start_line = wm_start_line + 1
+    wm_start_line = wm_start_line + 2
+    wm_end_line = wm_start_line
+    while gmx_output_lines[wm_end_line] != "\n":
+            wm_end_line = wm_end_line + 1
+    return gmx_output_lines[wm_start_line:wm_end_line]
 
 
 # Steps mark work as done.
@@ -1871,130 +1057,6 @@ def root_window(status, s_params, parent):
     set_variables(v1_name.get(), v2_group, v3_force, v4_water, water_v, check1_button, s_params)
 
 
-# Molecular Dynamics Performing window
-class CalculationWindow:
-    tasks_to_do = 0
-    bar_var = ""
-    bar_widget = ""
-    start_button = ""
-    stop_button = ""
-    log_button = ""
-
-    def __init__(self):
-        self.queue_status = Queue.Queue()
-        self.queue_percent = Queue.Queue()
-
-    # This will prevent Calculation Window to display if non protein has been selected
-    def check_window(self, master, g_parent, s_params, status):
-        project_name = s_params.project_name
-        if project_name != "nothing":
-            master.destroy()
-            root = Toplevel(g_parent)
-            self.window(root, s_params, status, g_parent)
-        elif project_name == "nothing":
-            no_molecule_warning()
-
-    # This function will create main Calculation Window
-    def window(self, root, s_params, status, parent):
-        root.wm_title("Calculation Window")
-        frame1 = Frame(root)
-        frame1.pack(side=TOP)
-        frame2 = Frame(root)
-        frame2.pack(side=TOP)
-
-        self.bar_var = StringVar(root)
-        self.bar_var.set("Ready to start")
-
-        w5 = Label(frame1, textvariable=self.bar_var)
-        w5.pack(side=TOP)
-        self.bar_widget = Progressbar(frame1)
-        self.bar_widget.pack(side=TOP)
-
-        exit_button = Button(frame2, text="EXIT", command=root.destroy)
-        exit_button.pack(side=LEFT)
-
-        save_button = Button(frame2, text="SAVE", command=lambda: select_file_save(1))
-        save_button.pack(side=LEFT)
-
-        stop_button = Button(frame2, text="STOP", command=lambda: self.start_counting(0))
-        stop_button.pack(side=LEFT)
-        stop = s_params.stop
-        if stop:
-            stop_button.configure(state=DISABLED)
-        self.stop_button = stop_button
-
-        start_button = Button(frame2, text="START", command=lambda: self.start_counting(1))
-        start_button.pack(side=LEFT)
-        if stop == 0:
-            start_button.configure(state=DISABLED)
-        self.start_button = start_button
-
-        log_button = Button(frame2, text="LOG", command=log_window)
-        log_button.pack(side=LEFT)
-        log_button.configure(state=DISABLED)
-        self.log_button = log_button
-
-        # Updateing status bar
-        tasks_nr = 0.0
-        for task in s_params.progress.to_do:
-            tasks_nr = tasks_nr + task
-        self.tasks_to_do = tasks_nr
-        thread.start_new_thread(self.bar_update, s_params, status)
-        self.bar_display(root, parent, s_params)
-
-    # This function will update status bar during molecular dynamics simulation (beware this is separate thread)
-    def bar_update(self, s_params, status):
-        percent = 0.0
-        while s_params.stop:
-            time.sleep(0.5)
-        while percent != 100:  # and error == ""
-            time.sleep(0.5)
-            percent = steps_status_bar("only_bar", s_params)
-            self.queue_percent.put(percent)
-            if s_params.stop == 0:
-                self.queue_status.put(status[1])
-            elif s_params.stop == 1:
-                self.queue_status.put("User Stoped")
-
-    #        if error != "":
-    #            self.queue_status.put("Fatal Error")
-
-    # This function will update status bar in thread safe manner
-    def bar_display(self, root, parent, s_params):
-        try:
-            status = self.queue_status.get(block=False)
-            self.bar_var.set(status)
-        except Queue.Empty:
-            status = "No change"
-        try:
-            percent = self.queue_percent.get(block=False)
-            self.bar_widget.configure(value=percent)
-        except:
-            pass
-        if status == "Fatal Error":
-            self.start_counting(0)
-            self.start_button.configure(state=DISABLED)
-            tkMessageBox.showerror("GROMACS Error Message", "Error")  # error)
-        if status == "Finished!":
-            root.destroy()
-            # Show interpretation window after successful completion of the calculations...
-            show_interpretation_window(parent, s_params)
-        else:
-            root.after(100, self.bar_display, root)
-
-    # This function will change global value if stop is clicked during simulation
-    def start_counting(self, value):
-        if value == 1:
-            stop = 0
-            thread.start_new_thread(dynamics, ())
-            self.stop_button.configure(state=ACTIVE)
-            self.start_button.configure(state=DISABLED)
-            self.log_button.configure(state=DISABLED)
-        elif value == 0:
-            stop = 1
-            self.stop_button.configure(state=DISABLED)
-            self.start_button.configure(state=ACTIVE)
-            self.log_button.configure(state=ACTIVE)
 
 
 # This window will allow to manipulate final molecule to interprate MD simulation results
